agent-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
langgraph==0.2.60
langchain-core==0.3.28
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
redis[hiredis]>=5.0
arq>=0.26
structlog>=24.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-instrumentation-fastapi>=0.41b0
pgvector>=0.3
tiktoken>=0.7
gitpython>=3.1
openai>=1.50
aiosmtplib>=3.0
pytest>=8.0
pytest-asyncio>=0.24
